fix Linux build

+++ src/xenia/kernel/xam/content_package.cc
@@ -119,7 +119,7 @@ X_RESULT FolderContentPackage::GetThumbnail(std::vector<uint8_t>* buffer) {
     if (map) {
       auto* header = (vfs::StfsHeader*)map->data();
       uint32_t thumb_size = std::min((uint32_t)header->metadata.thumbnail_size,
-                                     xe::vfs::XContentMetadata::kThumbLengthV1);
+                           (uint32_t)xe::vfs::XContentMetadata::kThumbLengthV1);
       buffer->resize(thumb_size);
       memcpy(buffer->data(), header->metadata.thumbnail, thumb_size);
       result = X_ERROR_SUCCESS;
@@ -173,7 +173,7 @@ X_RESULT StfsContentPackage::GetThumbnail(std::vector<uint8_t>* buffer) {
     return X_ERROR_DEVICE_NOT_CONNECTED;
   }
   uint32_t thumb_size = std::min((uint32_t)header_.metadata.thumbnail_size,
-                                 xe::vfs::XContentMetadata::kThumbLengthV1);
+                       (uint32_t)xe::vfs::XContentMetadata::kThumbLengthV1);
   buffer->resize(thumb_size);
   memcpy(buffer->data(), header_.metadata.thumbnail, thumb_size);
   return X_ERROR_SUCCESS;
